+++ x/sanction/module/module.go
@@ -77,7 +77,7 @@ func (a AppModuleBasic) GetQueryCmd() *cobra.Command {
 
 // GetTxCmd returns the transaction commands for the sanction module
 func (a AppModuleBasic) GetTxCmd() *cobra.Command {
-	return nil
+	return cli.TxCmd()
 }
 
 // RegisterGRPCGatewayRoutes registers the gRPC Gateway routes for the sanction module.